Refactored Cart Domain Structure into slices
