+++ Sample/ECommerce/Carts/Carts/Carts/Cart.cs
@@ -2,7 +2,10 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Ardalis.GuardClauses;
-using Carts.Carts.Events;
+using Carts.Carts.AddingProduct;
+using Carts.Carts.ConfirmingCart;
+using Carts.Carts.InitializingCart;
+using Carts.Carts.RemovingProduct;
 using Carts.Carts.ValueObjects;
 using Carts.Pricing;
 using Core.Aggregates;
