+++ Sample/ECommerce/Carts/Carts/Carts/GettingCartById/GetCartById.cs
@@ -1,0 +1,42 @@
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+using Ardalis.GuardClauses;
+using Core.Queries;
+using Marten;
+
+namespace Carts.Carts.GettingCartById
+{
+    public class GetCartById : IQuery<CartDetails>
+    {
+        public Guid CartId { get; }
+
+        private GetCartById(Guid cartId)
+        {
+            CartId = cartId;
+        }
+
+        public static GetCartById Create(Guid cartId)
+        {
+            Guard.Against.Default(cartId, nameof(cartId));
+
+            return new GetCartById(cartId);
+        }
+    }
+
+    internal class HandleGetCartById :
+        IQueryHandler<GetCartById, CartDetails?>
+    {
+        private readonly IDocumentSession querySession;
+
+        public HandleGetCartById(IDocumentSession querySession)
+        {
+            this.querySession = querySession;
+        }
+
+        public Task<CartDetails?> Handle(GetCartById request, CancellationToken cancellationToken)
+        {
+            return querySession.LoadAsync<CartDetails>(request.CartId, cancellationToken);
+        }
+    }
+}
