+++ Sample/ECommerce/Carts/Carts/Carts/GettingCarts/CartShortInfo.cs
@@ -1,8 +1,11 @@
 using System;
-using Carts.Carts.Events;
+using Carts.Carts.AddingProduct;
+using Carts.Carts.ConfirmingCart;
+using Carts.Carts.InitializingCart;
+using Carts.Carts.RemovingProduct;
 using Marten.Events.Aggregation;
 
-namespace Carts.Carts.Projections
+namespace Carts.Carts.GettingCarts
 {
     public class CartShortInfo
     {
