+++ Sample/ECommerce/Carts/Carts/Carts/InitializingCart/InitializeCart.cs
@@ -1,0 +1,53 @@
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+using Ardalis.GuardClauses;
+using Core.Commands;
+using Core.Repositories;
+using MediatR;
+
+namespace Carts.Carts.InitializingCart
+{
+    public class InitializeCart: ICommand
+    {
+        public Guid CartId { get; }
+
+        public Guid ClientId { get; }
+
+        private InitializeCart(Guid cartId, Guid clientId)
+        {
+            CartId = cartId;
+            ClientId = clientId;
+        }
+
+        public static InitializeCart Create(Guid cartId, Guid clientId)
+        {
+            Guard.Against.Default(cartId, nameof(cartId));
+            Guard.Against.Default(clientId, nameof(clientId));
+
+            return new InitializeCart(cartId, clientId);
+        }
+    }
+
+    internal class HandleInitializeCart:
+        ICommandHandler<InitializeCart>
+    {
+        private readonly IRepository<Cart> cartRepository;
+
+        public HandleInitializeCart(
+            IRepository<Cart> cartRepository
+        )
+        {
+            this.cartRepository = cartRepository;
+        }
+
+        public async Task<Unit> Handle(InitializeCart command, CancellationToken cancellationToken)
+        {
+            var cart = Cart.Initialize(command.CartId, command.ClientId);
+
+            await cartRepository.Add(cart, cancellationToken);
+
+            return Unit.Value;
+        }
+    }
+}
